+++ src/render/Integrator.cpp
@@ -188,6 +188,8 @@ namespace render {
     for (const auto& sample : samples) {
       State state;
       state.timeSample = sample.timeSample;
+      state.animationFrame = sample.animationFrame;
+      state.animationTime = sample.animationTime;
       state.sampleStream = sample.sampleStream();
       const Colord color = radiance(scene, sample.ray, state, recursiveRayCaster);
       if (metrics) {